+++ DutchiePay/src/main/java/dutchiepay/backend/domain/user/service/UserService.java
@@ -144,72 +144,6 @@ public class UserService {
         userRepository.save(user);
     }
 
-    @Transactional
-    public void unlinkKakao(UserDetailsImpl userDetails) {
-        OAuth2AuthorizedClient authorizedClient = oauthService.loadAuthorizedClient(
-            "kakao", // OAuth2 로그인 제공자 이름 (예: "google", "naver")
-            userDetails.getUsername() // 현재 인증된 사용자
-        );
-        String kakaoAccess = null;
-        if (authorizedClient != null) {
-            kakaoAccess = authorizedClient.getAccessToken().getTokenValue();// Access Token 추출
-        }
-        RestTemplate restTemplate = new RestTemplate();
-
-        // POST 요청으로 데이터 전송
-        // HttpHeaders 생성
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "Bearer " + kakaoAccess); // Authorization 헤더 설정
-
-        // HttpEntity에 본문 없이 헤더만 담기
-        HttpEntity<Void> entity = new HttpEntity<>(headers);
-
-        // POST 요청 보내기 (request body 없음)
-        restTemplate.exchange(
-            "https://kapi.kakao.com/v1/user/unlink",  // 요청할 URL
-            HttpMethod.POST,                 // HTTP 메서드
-            entity,                          // HttpEntity (본문 없음, 헤더만 있음)
-            String.class                     // 응답 타입
-        );
-    }
-
-    @Transactional
-    public void unlinkNaver(UserDetailsImpl userDetails) {
-
-        OAuth2AuthorizedClient authorizedClient = oauthService.loadAuthorizedClient(
-            "naver", // OAuth2 로그인 제공자 이름 (예: "google", "naver")
-            userDetails.getUsername() // 현재 인증된 사용자
-        );
-
-        String naverAccess = null;
-        if (authorizedClient != null) {
-            naverAccess = authorizedClient.getAccessToken().getTokenValue();// Access Token 추출
-        }
-        RestTemplate restTemplate = new RestTemplate();
-
-        // POST 요청으로 데이터 전송
-        String data = "?client_id=" + naverClientId +
-            "&client_secret=" + naverClientSecret +
-            "&access_token=" + naverAccess +
-            "&service_provider=NAVER" +
-            "&grant_type=delete";
-
-        restTemplate.exchange(
-            "https://nid.naver.com/oauth2.0/token" + data,  // 요청할 URL
-            HttpMethod.POST,                 // HTTP 메서드
-            null,                          // HttpEntity
-            String.class                     // 응답 타입
-        );
-    }
-
-    @Transactional
-    public void deleteOauthUser(HttpServletRequest request, UserDetailsImpl userDetails) {
-        userRepository.findByOauthProviderAndEmail(userDetails.getOAuthProvider(), userDetails.getEmail())
-                .orElseThrow(() -> new UserErrorException(UserErrorCode.USER_NOT_FOUND)).delete();
-        redisService.addBlackList(userDetails.getUserId(), request.getHeader("Authorization"));
-
-    }
-
     /**
      * 이메일 회원 탈퇴
      * AT는 BlackList 처리, RT는 삭제 처리
