+++ DutchiePay/src/main/java/dutchiepay/backend/global/oauth/handler/CustomAuthenticationException.java
@@ -1,0 +1,13 @@
+package dutchiepay.backend.global.oauth.handler;
+
+import lombok.Getter;
+import org.springframework.security.core.AuthenticationException;
+
+@Getter
+public class CustomAuthenticationException extends AuthenticationException {
+
+    public CustomAuthenticationException(String message) {
+        super(message);
+    }
+
+}
