+++ DutchiePay/src/main/java/dutchiepay/backend/global/oauth/handler/CustomOAuth2SuccessHandler.java
@@ -1,8 +1,6 @@
 package dutchiepay.backend.global.oauth.handler;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import dutchiepay.backend.domain.user.dto.UserLoginResponseDto;
 import dutchiepay.backend.domain.user.exception.UserErrorCode;
 import dutchiepay.backend.domain.user.exception.UserErrorException;
 import dutchiepay.backend.domain.user.repository.UserRepository;
@@ -10,23 +8,19 @@ import dutchiepay.backend.entity.User;
 import dutchiepay.backend.global.jwt.JwtUtil;
 import dutchiepay.backend.global.jwt.redis.RedisService;
 import dutchiepay.backend.global.oauth.dto.OAuthSuccessDto;
+import dutchiepay.backend.global.oauth.service.CustomOAuth2UserService;
 import dutchiepay.backend.global.security.UserDetailsImpl;
-import jakarta.annotation.PostConstruct;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientService;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
-import javax.crypto.Cipher;
-import javax.crypto.spec.SecretKeySpec;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 
 @Slf4j
 @Component
@@ -36,20 +30,8 @@ public class CustomOAuth2SuccessHandler implements AuthenticationSuccessHandler 
     private final JwtUtil jwtUtil;
     private final UserRepository userRepository;
     private final RedisService redisService;
-    private static String ENCRYPT_SECRET_KEY;
-    private static String ALGORITHM;
-
-    @Value("${ENCRYPT_SECRET_KEY}")
-    private String tempEncryptSecretKey;
-
-    @Value("${ALGORITHM}")
-    private String tempAlgorithm;
-
-    @PostConstruct
-    public void init() {
-        ENCRYPT_SECRET_KEY = tempEncryptSecretKey;
-        ALGORITHM = tempAlgorithm;
-    }
+    private final CustomOAuth2UserService customOAuth2UserService;
+    private final OAuth2AuthorizedClientService oauthService;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
@@ -58,6 +40,16 @@ public class CustomOAuth2SuccessHandler implements AuthenticationSuccessHandler 
         log.info("Oauth2 login success: User@{}", oAuth2User.getUserId());
 
         User user = userRepository.findById(oAuth2User.getUserId()).orElseThrow(() -> new UserErrorException(UserErrorCode.USER_NOT_FOUND));
+
+        if (user.getState() == 2) {
+            if (user.getOauthProvider().equals("kakao")) {
+                customOAuth2UserService.unlinkKakao(oAuth2User);
+            }
+            else {
+                customOAuth2UserService.unlinkNaver(oAuth2User);
+            }
+            throw new CustomAuthenticationException("해당하는 유저가 없습니다.");
+        }
 
         // refresh token 발급 후 저장
         String refreshToken = jwtUtil.createRefreshToken(user.getUserId());
@@ -78,7 +70,6 @@ public class CustomOAuth2SuccessHandler implements AuthenticationSuccessHandler 
         response.setContentType("application/json; charset=UTF-8");
         response.getWriter().write(new ObjectMapper().writeValueAsString(dto));
 
-        userRepository.save(user);
     }
 
 }
